[java] Move rule CloneMethodMustImplementCloneable from typeresolution to clone Replace existing rule with the typeresolution-based implementation.
