+++ docs/pages/release_notes.md
@@ -174,12 +174,18 @@ Notice this last scenario is slightly different to the Java syntax. This is due 
 
 *   The rule `LooseCoupling` (ruleset `java-coupling`) has been replaced by the typeresolution-based implementation.
 
+*   The rule `CloneMethodMustImplementCloneable` (ruleset `java-clone`) has been replaced by the typeresolution-based
+    implementation and is now able to detect cases if a class implements or extends a Cloneable class/interface.
+
 #### Deprecated Rules
 
 *   The rules `NcssConstructorCount`, `NcssMethodCount`, and `NcssTypeCount` (ruleset `java-codesize`) have been
     deprecated. They will be replaced by the new rule `NcssCount` in the same ruleset.
 
 *   The rule `LooseCoupling` in ruleset `java-typeresolution` is deprecated. Use the rule with the same name from ruleset `java-coupling` instead.
+
+*   The rule `CloneMethodMustImplementCloneable` in ruleset `java-typeresolution`is deprecated. Use the rule with
+    the same name from ruleset `java-clone` instead.
 
 #### Removed Rules
